[feat] 포인트 이벤트 구현

+++ src/test/java/com/study/ecommerce/DatabaseCleanup.java
@@ -3,6 +3,7 @@ package com.study.ecommerce;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
+import jakarta.persistence.Query;
 import org.springframework.beans.factory.InitializingBean;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
@@ -10,6 +11,7 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.StringUtils;
 
 
+import java.math.BigInteger;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -30,7 +32,7 @@ public class DatabaseCleanup implements InitializingBean {
     }
 
     private static String camelCaseToSnakeCase(String str) {
-        return str.replaceAll("([a-z])([A-Z]+)", "$1_$2").toLowerCase();
+        return str.replaceAll("([a-z])([A-Z])", "$1_$2").toLowerCase();
     }
 
     @Transactional
@@ -38,9 +40,20 @@ public class DatabaseCleanup implements InitializingBean {
         entityManager.flush();
         entityManager.createNativeQuery("SET REFERENTIAL_INTEGRITY FALSE").executeUpdate();
         for (String tableName : tableNames) {
+            if (!isTableExists(tableName)) {
+                continue;
+            }
             entityManager.createNativeQuery("TRUNCATE TABLE " + tableName).executeUpdate();
             entityManager.createNativeQuery("ALTER TABLE " + tableName + " ALTER COLUMN ID RESTART WITH 1").executeUpdate();
         }
         entityManager.createNativeQuery("SET REFERENTIAL_INTEGRITY TRUE").executeUpdate();
+    }
+
+    private boolean isTableExists(String tableName) {
+        String checkTableQuery = "SELECT COUNT(*) FROM INFORMATION_SCHEMA.TABLES WHERE TABLE_NAME = :tableName";
+        Query query = entityManager.createNativeQuery(checkTableQuery);
+        query.setParameter("tableName", tableName.toUpperCase());
+        Long count = (Long) query.getSingleResult();
+        return count > 0;
     }
 }
